Skill-Update: skip skill=agent-capabilities reason="mainline refresh does not change particle capability guidance"

+++ core/audio/src/wavetable_authoring.cpp
@@ -388,10 +388,6 @@ WavetableCompileResult compile_wavetable(BufferView<const float> source, double 
         for (float& sample : base)
             sample *= gain;
 
-        // Authoring is offline and its materialized hash is part of the public
-        // content-addressing contract. Use the deterministic scalar backend on
-        // every platform instead of the float specialization, whose Apple vDSP
-        // path can produce different low-order samples across identical calls.
         signal::FftT<double> fft(static_cast<int>(recipe.table_length));
         if (!fft.ready()) {
             result.status = WavetableCompileStatus::FftUnavailable;
@@ -399,11 +395,11 @@ WavetableCompileResult compile_wavetable(BufferView<const float> source, double 
         }
         std::vector<std::complex<double>> spectrum(recipe.table_length);
         for (std::size_t i = 0; i < base.size(); ++i)
-            spectrum[i] = {base[i], 0.0f};
+            spectrum[i] = {static_cast<double>(base[i]), 0.0};
         fft.forward(spectrum.data());
         // DC removal is an output invariant even if finite sinc arithmetic left
         // a tiny residual after the source-cycle mean subtraction.
-        spectrum[0] = {0.0f, 0.0f};
+        spectrum[0] = {0.0, 0.0};
 
         const auto ceilings = signal::detail::build_wavetable_band_ceilings<float>(
             recipe.num_bands, static_cast<float>(recipe.reference_sample_rate));
@@ -425,7 +421,7 @@ WavetableCompileResult compile_wavetable(BufferView<const float> source, double 
             auto band_spectrum = spectrum;
             for (std::size_t k = maximum_harmonic + 1; k < band_spectrum.size() - maximum_harmonic;
                  ++k)
-                band_spectrum[k] = {0.0f, 0.0f};
+                band_spectrum[k] = {0.0, 0.0};
             fft.inverse(band_spectrum.data());
             signal::WavetableEntry band;
             band.max_frequency_hz = ceiling;
